refactor: removed goreleaserlib: moved all to main
